refactor: rename cell_executor/cell_launcher to executor/launcher
Drops the underscore-prefix. The two files are wiring at the bin level
sitting next to host.rs, runtime.rs, dispatcher/, ipfs.rs, etc. -- all
unprefixed. The cell_ prefix was redundant and read awkwardly; the
file-tree context already supplies the meaning.

Avoids a future name collision with crates/cell/ once Phase 1 lands.

Pure rename. cargo check --lib --bins --tests -p ww passes clean.
