+++ tests/runtime_spike_test.rs
@@ -15,7 +15,7 @@ use std::time::Instant;
 
 use tokio::sync::mpsc;
 
-use ww::cell_launcher::create_runtime_client;
+use ww::launcher::create_runtime_client;
 use ww::rpc::{CachePolicy, NetworkState};
 use ww::system_capnp;
