+++ tests/snap_hello_rs_http_listener_e2e.rs
@@ -18,8 +18,8 @@
 
 use tokio::sync::{mpsc, oneshot, watch};
 
-use ww::cell_launcher::create_runtime_client;
 use ww::dispatcher::server::{new_registry, CgiRequest, CgiResponse};
+use ww::launcher::create_runtime_client;
 use ww::rpc::{CachePolicy, NetworkState};
 use ww::system_capnp;
